+++ cmd/telegraf/controller.go
@@ -2,19 +2,32 @@ package main
 
 import (
 	"os"
+	"sync"
 )
 
 type PushController struct {
 	pushControllerChan chan bool
 	pushModuleChan     chan bool
+	pushModuleStopChan chan bool
 	signals            chan os.Signal
+	wg                 sync.WaitGroup
+	inputFilters       []string
+	outputFilters      []string
+	aggregatorFilters  []string
+	processorFilters   []string
 }
 
-func NewAgentPushController(pushControllerChan chan bool, pushModuleChan chan bool, signals chan os.Signal) PushController {
+func NewAgentPushController(pushControllerChan chan bool, pushModuleChan chan bool, signals chan os.Signal, inputFilters, outputFilters, aggregatorFilters, processorFilters []string) PushController {
+	pushModuleStopChan := make(chan bool, 1)
 	return PushController{
 		pushControllerChan: pushControllerChan,
 		pushModuleChan:     pushModuleChan,
+		pushModuleStopChan: pushModuleStopChan,
 		signals:            signals,
+		inputFilters:       inputFilters,
+		outputFilters:      outputFilters,
+		aggregatorFilters:  aggregatorFilters,
+		processorFilters:   processorFilters,
 	}
 }
 
@@ -22,21 +35,23 @@ func (pushController *PushController) StartPushController() {
 	for {
 		select {
 		case controllerTrigger := <-pushController.pushControllerChan:
-			switch controllerTrigger {
-			case false:
+			if !controllerTrigger {
 				pushController.pushModuleChan <- false
-				break
-			case true:
+			} else {
 				pushController.pushModuleChan <- true
 				go func() {
-					select {
-					case pushModuleTrigger := <-pushController.pushModuleChan:
-						if pushModuleTrigger {
-							pushController.startPushMonitoring()
+					for {
+						select {
+						case pushModuleTrigger := <-pushController.pushModuleChan:
+							if pushModuleTrigger {
+								go pushController.run()
+							} else {
+								pushController.pushModuleStopChan <- false
+								return
+							}
 						}
 					}
 				}()
-				break
 			}
 		}
 	}
