+++ cmd/telegraf/telegraf.go
@@ -1,52 +1,255 @@
 package main
 
 import (
-	"context"
+	"flag"
+	"fmt"
 	"github.com/influxdata/telegraf/cloudbarista/listener"
+	"github.com/influxdata/telegraf/config"
+	"github.com/influxdata/telegraf/internal"
+	"github.com/influxdata/telegraf/internal/goplugin"
+	"github.com/influxdata/telegraf/logger"
+	"github.com/influxdata/telegraf/plugins/inputs"
+	"github.com/influxdata/telegraf/plugins/outputs"
 	"log"
+	"net/http"
 	"os"
 	"os/signal"
+	"sort"
+	"strings"
 	"sync"
 	"syscall"
 )
 
-var wg sync.WaitGroup
-var CB_Push_Controller PushController
+// If you update these, update usage.go and usage_windows.go
+var fDebug = flag.Bool("debug", false,
+	"turn on debug logging")
+var pprofAddr = flag.String("pprof-addr", "",
+	"pprof address to listen on, not activate pprof if empty")
+var fQuiet = flag.Bool("quiet", false,
+	"run in quiet mode")
+var fTest = flag.Bool("test", false, "enable test mode: gather metrics, print them out, and exit. Note: Test mode only runs inputs, not processors, aggregators, or outputs")
+var fTestWait = flag.Int("test-wait", 0, "wait up to this many seconds for service inputs to complete in test mode")
+var fConfig = flag.String("config", "", "configuration file to load")
+var fConfigDirectory = flag.String("config-directory", "",
+	"directory containing additional *.conf files")
+var fVersion = flag.Bool("version", false, "display the version and exit")
+var fSampleConfig = flag.Bool("sample-config", false,
+	"print out full sample configuration")
+var fPidfile = flag.String("pidfile", "", "file to write our pid to")
+var fSectionFilters = flag.String("section-filter", "",
+	"filter the sections to print, separator is ':'. Valid values are 'agent', 'global_tags', 'outputs', 'processors', 'aggregators' and 'inputs'")
+var fInputFilters = flag.String("input-filter", "",
+	"filter the inputs to enable, separator is :")
+var fInputList = flag.Bool("input-list", false,
+	"print available input plugins.")
+var fOutputFilters = flag.String("output-filter", "",
+	"filter the outputs to enable, separator is :")
+var fOutputList = flag.Bool("output-list", false,
+	"print available output plugins.")
+var fAggregatorFilters = flag.String("aggregator-filter", "",
+	"filter the aggregators to enable, separator is :")
+var fProcessorFilters = flag.String("processor-filter", "",
+	"filter the processors to enable, separator is :")
+var fUsage = flag.String("usage", "",
+	"print usage for a plugin, ie, 'telegraf --usage mysql'")
+var fService = flag.String("service", "",
+	"operate on the service (windows only)")
+var fServiceName = flag.String("service-name", "telegraf", "service name (windows only)")
+var fServiceDisplayName = flag.String("service-display-name", "Telegraf Data Collector Service", "service display name (windows only)")
+var fRunAsConsole = flag.Bool("console", false, "run as console application (windows only)")
+var fPlugins = flag.String("plugin-directory", "",
+	"path to directory containing external plugins")
+var fRunOnce = flag.Bool("once", false, "run one gather and exit")
+
+var (
+	version string
+	commit  string
+	branch  string
+)
+
+func usageExit(rc int) {
+	fmt.Println(internal.Usage)
+	os.Exit(rc)
+}
 
 func main() {
+	var wg sync.WaitGroup
+
+	pullListenerSignals := make(chan os.Signal, 100)
+	pushModuleSignals := make(chan os.Signal, 100)
 	pushControllerChan := make(chan bool, 1)
 	pushModuleChan := make(chan bool, 1)
-	signals := make(chan os.Signal, 100)
 
-	ctx, cancel := context.WithCancel(context.Background())
-	CB_Listener := listener.NewAgentPullListener(8888, pushControllerChan, signals, ctx, cancel)
+	flag.Usage = func() { usageExit(0) }
+	flag.Parse()
+	args := flag.Args()
 
-	wg.Add(1)
+	sectionFilters, inputFilters, outputFilters := []string{}, []string{}, []string{}
+	if *fSectionFilters != "" {
+		sectionFilters = strings.Split(":"+strings.TrimSpace(*fSectionFilters)+":", ":")
+	}
+	if *fInputFilters != "" {
+		inputFilters = strings.Split(":"+strings.TrimSpace(*fInputFilters)+":", ":")
+	}
+	if *fOutputFilters != "" {
+		outputFilters = strings.Split(":"+strings.TrimSpace(*fOutputFilters)+":", ":")
+	}
+
+	aggregatorFilters, processorFilters := []string{}, []string{}
+	if *fAggregatorFilters != "" {
+		aggregatorFilters = strings.Split(":"+strings.TrimSpace(*fAggregatorFilters)+":", ":")
+	}
+	if *fProcessorFilters != "" {
+		processorFilters = strings.Split(":"+strings.TrimSpace(*fProcessorFilters)+":", ":")
+	}
+
+	logger.SetupLogging(logger.LogConfig{})
+
+	// Load external plugins, if requested.
+	if *fPlugins != "" {
+		log.Printf("I! Loading external plugins from: %s", *fPlugins)
+		if err := goplugin.LoadExternalPlugins(*fPlugins); err != nil {
+			log.Fatal("E! " + err.Error())
+		}
+	}
+
+	if *pprofAddr != "" {
+		go func() {
+			pprofHostPort := *pprofAddr
+			parts := strings.Split(pprofHostPort, ":")
+			if len(parts) == 2 && parts[0] == "" {
+				pprofHostPort = fmt.Sprintf("localhost:%s", parts[1])
+			}
+			pprofHostPort = "http://" + pprofHostPort + "/debug/pprof"
+
+			log.Printf("I! Starting pprof HTTP server at: %s", pprofHostPort)
+
+			if err := http.ListenAndServe(*pprofAddr, nil); err != nil {
+				log.Fatal("E! " + err.Error())
+			}
+		}()
+	}
+
+	if len(args) > 0 {
+		switch args[0] {
+		case "version":
+			fmt.Println(formatFullVersion())
+			return
+		case "config":
+			config.PrintSampleConfig(
+				sectionFilters,
+				inputFilters,
+				outputFilters,
+				aggregatorFilters,
+				processorFilters,
+			)
+			return
+		}
+	}
+
+	// switch for flags which just do something and exit immediately
+	switch {
+	case *fOutputList:
+		fmt.Println("Available Output Plugins: ")
+		names := make([]string, 0, len(outputs.Outputs))
+		for k := range outputs.Outputs {
+			names = append(names, k)
+		}
+		sort.Strings(names)
+		for _, k := range names {
+			fmt.Printf("  %s\n", k)
+		}
+		return
+	case *fInputList:
+		fmt.Println("Available Input Plugins:")
+		names := make([]string, 0, len(inputs.Inputs))
+		for k := range inputs.Inputs {
+			names = append(names, k)
+		}
+		sort.Strings(names)
+		for _, k := range names {
+			fmt.Printf("  %s\n", k)
+		}
+		return
+	case *fVersion:
+		fmt.Println(formatFullVersion())
+		return
+	case *fSampleConfig:
+		config.PrintSampleConfig(
+			sectionFilters,
+			inputFilters,
+			outputFilters,
+			aggregatorFilters,
+			processorFilters,
+		)
+		return
+	case *fUsage != "":
+		err := config.PrintInputConfig(*fUsage)
+		err2 := config.PrintOutputConfig(*fUsage)
+		if err != nil && err2 != nil {
+			log.Fatalf("E! %s and %s", err, err2)
+		}
+		return
+	}
+
+	shortVersion := version
+	if shortVersion == "" {
+		shortVersion = "unknown"
+	}
+
+	// Configure version
+	if err := internal.SetVersion(shortVersion); err != nil {
+		log.Println("Telegraf version already configured to: " + internal.Version())
+	}
+
 	go func() {
-		signal.Notify(signals, os.Interrupt, syscall.SIGHUP,
+		signal.Notify(pullListenerSignals, os.Interrupt, syscall.SIGHUP,
 			syscall.SIGTERM, syscall.SIGINT)
-		<-signals
-		<-signals
-		wg.Done()
+		<-pullListenerSignals
+		return
+	}()
+	go func() {
+		signal.Notify(pushModuleSignals, os.Interrupt, syscall.SIGHUP,
+			syscall.SIGTERM, syscall.SIGINT)
+		<-pushModuleSignals
+		//pushControllerChan <- false
 		return
 	}()
 
-	CB_Push_Controller = NewAgentPushController(pushControllerChan, pushModuleChan, signals)
-	wg.Add(1)
+	CB_Push_Controller := NewAgentPushController(pushControllerChan, pushModuleChan, pushModuleSignals, inputFilters, outputFilters, aggregatorFilters, processorFilters)
+	CB_Listener := listener.NewAgentPullListener(8888, pushControllerChan, pullListenerSignals)
 	go func() {
-		err := CB_Listener.Start()
-		if err != nil {
-			log.Fatalf("Running Cloud-Barista API Server Error")
-		}
-		wg.Done()
+		_ = CB_Listener.Start()
 	}()
 
 	wg.Add(1)
 	go func() {
 		CB_Push_Controller.StartPushController()
 		//push.StartPushController(pushControllerChan, pushModuleChan, signals)
-		wg.Done()
 	}()
 	wg.Wait()
 	return
+}
+
+func formatFullVersion() string {
+	var parts = []string{"Telegraf"}
+
+	if version != "" {
+		parts = append(parts, version)
+	} else {
+		parts = append(parts, "unknown")
+	}
+
+	if branch != "" || commit != "" {
+		if branch == "" {
+			branch = "unknown"
+		}
+		if commit == "" {
+			commit = "unknown"
+		}
+		git := fmt.Sprintf("(git: %s %s)", branch, commit)
+		parts = append(parts, git)
+	}
+
+	return strings.Join(parts, " ")
 }
